+++ src/mopipe/core/common/util.py
