+++ src/mopipe/core/data/empirical.py
@@ -4,10 +4,10 @@ import typing as t
 
 from pandas import DataFrame, Series
 
-from mopipe.common import MocapMetadataEntries, maybe_generate_id
+from mopipe.core.common import MocapMetadataEntries, maybe_generate_id
 
 if t.TYPE_CHECKING:
-    from mopipe.data import ExperimentLevel
+    from mopipe.core.data import ExperimentLevel
 
 
 class MetaData(dict):
